backend/requirements/base.txt
Django>=4.2,<5

channels

djangorestframework

djangorestframework-simplejwt

django-tenants

django-cors-headers

python-dotenv

dj-database-url

psycopg2-binary
